fix lines
